refactor(data): promote aliasesProp out of internals into @/data/properties
aliasesProp lived in src/data/internals/coreProperties.ts but had zero
internal-kernel importers — every consumer (src/data/* and ~9 plugins)
reached across the internals boundary for it. It's a public property
descriptor that was just parked in internals.

Move the definition into @/data/properties (alongside typesProp et al.,
where it was already re-exported as the canonical public path), delete the
internals leaf, and repoint all importers to @/data/properties. The leaf's
cycle-avoidance rationale was already moot: blockTypes/stateBlocks/targets
all import @/data/properties anyway, and properties.ts now imports nothing
from internals.

Part of #139 (promotion-first pass; no lint rule).
