+++ src/plugins/alias/collisionMerge.ts
@@ -4,7 +4,7 @@ import {
   defineMutator,
   type BlockData,
 } from '@/data/api'
-import { aliasesProp } from '@/data/internals/coreProperties'
+import { aliasesProp } from '@/data/properties'
 import { mergeBlocksInTx } from '@/data/internals/blockMerge'
 import { mergeProperties } from '@/data/internals/mergeProperties'
