+++ src/plugins/alias/test/collisionMerge.test.ts
@@ -5,7 +5,7 @@ import { ChangeScope, type BlockData } from '@/data/api'
 import { BlockCache } from '@/data/blockCache'
 import { createTestDb, resetTestDb, type TestDb } from '@/data/test/createTestDb'
 import { Repo } from '@/data/repo'
-import { aliasesProp } from '@/data/internals/coreProperties'
+import { aliasesProp } from '@/data/properties'
 import { kernelDataExtension } from '@/data/kernelDataExtension.js'
 import { resolveFacetRuntimeSync } from '@/facets/facet.js'
 import { aliasDataExtension } from '../dataExtension.ts'
