+++ src/plugins/geo/LocationPropertyEditor.tsx
@@ -25,8 +25,7 @@ import { createOrFindPlaceInteractive } from './placeNameCollision'
 import { pickCurrentLocation } from './pickCurrentLocation'
 import { resolveApiKey } from './googlePlacesClient'
 import { usePlaceSearch, type PlaceSearchResult } from './usePlaceSearch'
-import { typesProp } from '@/data/properties'
-import { aliasesProp } from '@/data/internals/coreProperties'
+import { aliasesProp, typesProp } from '@/data/properties'
 import { PLACE_TYPE } from './blockTypes'
 import { placeAddressProp } from './properties'
