+++ src/plugins/geo/createOrFindPlace.ts
@@ -25,8 +25,7 @@ import { ChangeScope, type BlockData, type Tx } from '@/data/api'
 import type { Block } from '@/data/block'
 import type { Repo } from '@/data/repo'
 import { PAGE_TYPE } from '@/data/blockTypes'
-import { aliasesProp } from '@/data/internals/coreProperties'
-import { typesProp } from '@/data/properties'
+import { aliasesProp, typesProp } from '@/data/properties'
 import { keyAtEnd } from '@/data/orderKey'
 import { PLACE_TYPE } from './blockTypes'
 import {
