+++ src/plugins/references/test/mergeRetargetProcessor.test.ts
@@ -5,7 +5,7 @@ import { ChangeScope, normalizeReferences, type BlockData } from '@/data/api'
 import { BlockCache } from '@/data/blockCache'
 import { kernelDataExtension } from '@/data/kernelDataExtension.js'
 import { Repo } from '@/data/repo'
-import { aliasesProp } from '@/data/internals/coreProperties'
+import { aliasesProp } from '@/data/properties'
 import { createTestDb, resetTestDb, type TestDb } from '@/data/test/createTestDb'
 import { resolveFacetRuntimeSync } from '@/facets/facet.js'
 import { aliasDataExtension } from '@/plugins/alias/dataExtension.js'
